// src/app/core/models/pokemon.model.ts
export interface Pokemon {
    id: number;
    name: string;
    // base_experience: number;
    // height: number;
    size: number;
    // is_default: boolean;
    // order: number;
    // weight: number;
    imageUrl: string;
    abilities: Ability[];
    // forms: Form[];
    // game_indices: Index[];
    // held_items: HeldItem[];
    // location_area_encounters: string;
    // moves: Mfe[];
    // species: Species;
    // sprites: Sprites;
    // cries: Cries;
    // stats: Stat[];
    types: string[];
    // past_types: PastType[];
  }
  
  export interface Ability {
    is_hidden: boolean;
    slot: number;
    ability: Ability2;
  }
  
  export interface Ability2 {
    name: string;
    url: string;
  }